chore(Components): move playwright to root [YTFRONT-5601]
