+++ packages/components/src/components/ClipboardButton/__tests__/ClipboardButton.visual.test.tsx
@@ -1,4 +1,4 @@
-import {test} from '../../../playwright-components/core';
+import {test} from '../../../../playwright-components/core';
 
 import {ClipboardButton} from '../ClipboardButton';
