+++ packages/components/src/components/MetaTable/__tests__/MetaTable.visual.test.tsx
@@ -1,4 +1,4 @@
-import {test} from '../../../playwright-components/core';
+import {test} from '../../../../playwright-components/core';
 
 import {MetaTable} from '../MetaTable';
 import {
