+++ packages/components/src/components/NoContent/__tests__/NoContent.visual.test.tsx
@@ -1,4 +1,4 @@
-import {test} from '../../../playwright-components/core';
+import {test} from '../../../../playwright-components/core';
 
 import {NoContent} from '../NoContent';
