+++ packages/components/src/components/SchemaDataType/DataType/__tests__/DataType.visual.test.tsx
@@ -1,6 +1,6 @@
 import {type ReactElement} from 'react';
 
-import {test} from '../../../../playwright-components/core';
+import {test} from '../../../../../playwright-components/core';
 
 import {DataType} from '../DataType';
 import {
