+++ packages/components/src/components/Tooltip/__tests__/Tooltip.visual.test.tsx
@@ -1,4 +1,4 @@
-import {test} from '../../../playwright-components/core';
+import {test} from '../../../../playwright-components/core';
 
 import {Tooltip} from '../Tooltip';
