+++ packages/components/src/components/YqlValue/__tests__/YqlValue.visual.test.tsx
@@ -1,4 +1,4 @@
-import {test} from '../../../playwright-components/core';
+import {test} from '../../../../playwright-components/core';
 
 import {YqlValue} from '../YqlValue';
 import {
