+++ packages/components/tsconfig.publish.json
@@ -6,5 +6,6 @@
     "emitDeclarationOnly": true,
     "noEmit": false
   },
-  "include": ["src/**/*"]
+  "include": ["src/**/*"],
+  "exclude": ["node_modules", "build", "**/*.visual.test.tsx"]
 }
